video: driver: fix list usage without init
If workqueue creation fails, driver is accessing buffer lists
in session close sequence, which is not yet inited and leading
to NULL ptr dereferrence issue, while traversing uninitialized
list. So added change to avoid accessing list without init
completion.

+++ driver/vidc/inc/msm_vidc_internal.h
@@ -828,7 +828,7 @@ struct msm_vidc_timestamp {
 struct msm_vidc_timestamps {
 	struct list_head       list;
 	u32                    count;
-	u32                    rank;
+	u64                    rank;
 };
 
 struct msm_vidc_pool {
